% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{planilha_}
\alias{planilha_}
\title{Exporta uma planilha de dados}
\usage{
planilha_()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função exporta uma planilha genérica
}
