% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faturamento.R
\name{planilha_faturamento}
\alias{planilha_faturamento}
\title{Exporta a planilha de faturamento do Fluxo Ágil}
\usage{
planilha_faturamento()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função exporta a planilha de faturamento do Fluxo Ágil
}
