% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoramento.R
\name{planilha_monitoramento}
\alias{planilha_monitoramento}
\title{Exporta a planilha de monitoramento do Fluxo Ágil}
\usage{
planilha_monitoramento()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função exporta a planilha de faturamento do Fluxo Ágil
}
