% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renova-package.R
\docType{package}
\name{Renova-package}
\alias{Renova}
\alias{Renova-package}
\title{Renova: Cria planilhas para o Fluxo Ágil}
\description{
Este pacote acessa a API "Renova_BI" e gera planilhas para relatórios do Fluxo Ágil.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HerkenhoffPrates/Renova}
  \item Report bugs at \url{https://github.com/HerkenhoffPrates/Renova/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
