% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{renova_VPN}
\alias{renova_VPN}
\title{Conecta a VPN da Renova via 'ODBC Driver 18 for SQL Server' para acessar a base de dados 'RENOVA_POWER_BI'}
\usage{
renova_VPN()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Esta função faz a conexão entre a VPN da Renova e o R via 'ODBC Driver 18 for SQL Server' para acessar a base de dados 'RENOVA_POWER_BI'
}
\note{
Baixe e instale o driver "ODBC Driver 18 for SQL Server" no site da Microsoft.

Veja a lista de drivers instalados na sua máquina com o comando 'odbc::odbcListDrivers()'
}
